# internal/generated/api.gen.go
//go:build go1.22

// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

// BusDisplayInfo defines model for BusDisplayInfo.
type BusDisplayInfo struct {
	// IsWheelchair Whether the bus has wheelchair accessibility
	IsWheelchair bool `json:"IsWheelchair"`

	// LoadStatus Array of load statuses
	LoadStatus []string `json:"LoadStatus"`

	// NextBuses Array of arrival times in minutes
	NextBuses []string `json:"NextBuses"`

	// Operator Bus operator
	Operator string `json:"Operator"`

	// ServiceNo Bus service number
	ServiceNo string `json:"ServiceNo"`
}

// NearestBusStopWithArrivals defines model for NearestBusStopWithArrivals.
type NearestBusStopWithArrivals struct {
	// Arrivals Array of bus arrival information for this bus stop
	Arrivals []BusDisplayInfo `json:"Arrivals"`

	// BusStopCode Bus stop code
	BusStopCode string `json:"BusStopCode"`

	// Description Description of the bus stop location
	Description string `json:"Description"`

	// Distance Distance from the provided coordinates in kilometers
	Distance float32 `json:"Distance"`

	// RoadName Road name where the bus stop is located
	RoadName string `json:"RoadName"`
}

// NearestBusStopsResponse Array of the nearest 4 bus stops with their arrival information, sorted by nearest first
type NearestBusStopsResponse = []NearestBusStopWithArrivals

// GetFakeDataParams defines parameters for GetFakeData.
type GetFakeDataParams struct {
	// Name Name parameter
	Name *string `form:"name,omitempty" json:"name,omitempty"`
}

// GetBusArrivalParams defines parameters for GetBusArrival.
type GetBusArrivalParams struct {
	// BusStopCode Bus stop code
	BusStopCode string `form:"busStopCode" json:"busStopCode"`
}

// GetNearestBusStopsParams defines parameters for GetNearestBusStops.
type GetNearestBusStopsParams struct {
	// Lat Latitude coordinate
	Lat float32 `form:"lat" json:"lat"`

	// Lng Longitude coordinate
	Lng float32 `form:"lng" json:"lng"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get fake data
	// (GET /api)
	GetFakeData(w http.ResponseWriter, r *http.Request, params GetFakeDataParams)
	// Get bus arrival information
	// (GET /getBusArrival)
	GetBusArrival(w http.ResponseWriter, r *http.Request, params GetBusArrivalParams)
	// Get nearest bus stops with arrivals
	// (GET /getNearestBusStops)
	GetNearestBusStops(w http.ResponseWriter, r *http.Request, params GetNearestBusStopsParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetFakeData operation middleware
func (siw *ServerInterfaceWrapper) GetFakeData(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetFakeDataParams

	// ------------- Optional query parameter "name" -------------

	err = runtime.BindQueryParameter("form", true, false, "name", r.URL.Query(), &params.Name)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "name", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetFakeData(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetBusArrival operation middleware
func (siw *ServerInterfaceWrapper) GetBusArrival(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetBusArrivalParams

	// ------------- Required query parameter "busStopCode" -------------

	if paramValue := r.URL.Query().Get("busStopCode"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "busStopCode"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "busStopCode", r.URL.Query(), &params.BusStopCode)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "busStopCode", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBusArrival(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetNearestBusStops operation middleware
func (siw *ServerInterfaceWrapper) GetNearestBusStops(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetNearestBusStopsParams

	// ------------- Required query parameter "lat" -------------

	if paramValue := r.URL.Query().Get("lat"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "lat"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "lat", r.URL.Query(), &params.Lat)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "lat", Err: err})
		return
	}

	// ------------- Required query parameter "lng" -------------

	if paramValue := r.URL.Query().Get("lng"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "lng"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "lng", r.URL.Query(), &params.Lng)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "lng", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetNearestBusStops(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/api", wrapper.GetFakeData)
	m.HandleFunc("GET "+options.BaseURL+"/getBusArrival", wrapper.GetBusArrival)
	m.HandleFunc("GET "+options.BaseURL+"/getNearestBusStops", wrapper.GetNearestBusStops)

	return m
}