# openapi.yaml
openapi: 3.0.3
info:
  title: When Bus API
  description: API for bus arrival information
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api:
    get:
      summary: Get fake data
      description: Returns fake data with an optional name parameter
      operationId: getFakeData
      parameters:
        - name: name
          in: query
          description: Name parameter
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
                example: "wow this is fake data!, u wrote John in the request!"

  /getBusArrival:
    get:
      summary: Get bus arrival information
      description: Returns bus arrival times for a given bus stop code
      operationId: getBusArrival
      parameters:
        - name: busStopCode
          in: query
          description: Bus stop code
          required: true
          schema:
            type: string
            example: "71119"
      responses:
        '200':
          description: Successful response with bus arrival information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusDisplayInfo'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string

  /getNearestBusStops:
    get:
      summary: Get nearest bus stops with arrivals
      description: Returns the nearest 4 bus stops with their bus arrival times, sorted by nearest first
      operationId: getNearestBusStops
      parameters:
        - name: lat
          in: query
          description: Latitude coordinate
          required: true
          schema:
            type: number
            format: float
            example: 1.326277
        - name: lng
          in: query
          description: Longitude coordinate
          required: true
          schema:
            type: number
            format: float
            example: 103.890342
      responses:
        '200':
          description: Successful response with nearest bus stops and their arrival times
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearestBusStopsResponse'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string

components:
  schemas:
    BusDisplayInfo:
      type: object
      properties:
        ServiceNo:
          type: string
          description: Bus service number
          example: "123"
        Operator:
          type: string
          description: Bus operator
          example: "SBST"
        NextBuses:
          type: array
          items:
            type: string
          description: Array of arrival times in minutes
          example: ["5", "12", "25"]
        LoadStatus:
          type: array
          items:
            type: string
          description: Array of load statuses
          example: ["SEA", "SDA", "LSD"]
        IsWheelchair:
          type: boolean
          description: Whether the bus has wheelchair accessibility
          example: true
      required:
        - ServiceNo
        - Operator
        - NextBuses
        - LoadStatus
        - IsWheelchair
    NearestBusStopsResponse:
      type: array
      items:
        $ref: '#/components/schemas/NearestBusStopWithArrivals'
      description: Array of the nearest 4 bus stops with their arrival information, sorted by nearest first
    NearestBusStopWithArrivals:
      type: object
      properties:
        BusStopCode:
          type: string
          description: Bus stop code
          example: "71119"
        RoadName:
          type: string
          description: Road name where the bus stop is located
          example: "Jurong West Ave 1"
        Description:
          type: string
          description: Description of the bus stop location
          example: "Blk 123"
        Distance:
          type: number
          format: float
          description: Distance from the provided coordinates in kilometers
          example: 0.15
        Arrivals:
          type: array
          items:
            $ref: '#/components/schemas/BusDisplayInfo'
          description: Array of bus arrival information for this bus stop
      required:
        - BusStopCode
        - RoadName
        - Description
        - Distance
        - Arrivals